pull

+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/util/TestMergingPageIterator.java
@@ -19,7 +19,7 @@ import io.trino.spi.PageBuilder;
 import io.trino.spi.connector.SortOrder;
 import io.trino.spi.type.Type;
 import io.trino.spi.type.TypeOperators;
-import org.testng.annotations.Test;
+import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
 import java.util.Iterator;
@@ -80,7 +80,7 @@ public class TestMergingPageIterator
                 .collect(toList());
         Iterator<Page> iterator = new MergingPageIterator(pages, types, sortIndexes, sortOrders, new TypeOperators());
 
-        List<Long> values = new ArrayList<>();
+        List<Integer> values = new ArrayList<>();
         while (iterator.hasNext()) {
             Page page = iterator.next();
             for (int i = 0; i < page.getPositionCount(); i++) {
@@ -89,8 +89,8 @@ public class TestMergingPageIterator
                     values.add(null);
                 }
                 else {
-                    long x = INTEGER.getLong(page.getBlock(0), i);
-                    long y = INTEGER.getLong(page.getBlock(1), i);
+                    int x = INTEGER.getInt(page.getBlock(0), i);
+                    int y = INTEGER.getInt(page.getBlock(1), i);
                     assertEquals(y, x * 22);
                     values.add(x);
                 }
